requirements.txt
numpy~=1.24.1
MiniSom~=2.3.1
scikit-learn~=1.3.0
tqdm~=4.65.0
pandas~=2.0.3
matplotlib~=3.7.2
statsmodels~=0.14.0
DateTime~=4.5
Pillow~=10.0.0
ipywidgets~=8.1.0
pydot~=1.4.2
ipython~=8.15.0
configparser~=5.3.0
wheel~=0.41.1
Jinja2~=3.1.2
rsa~=4.9
pyasn1~=0.5.0
comm~=0.1.4
traitlets~=5.9.0
gast~=0.4.0
h5py~=3.9.0
parso~=0.8.3
jedi~=0.19.0
colorama~=0.4.5
pytz~=2022.2.1
requests~=2.28.1
keras~=2.10.0
PyYAML~=6.0
tensorflow~=2.10.1
six~=1.16.0
scipy~=1.11.1
wrapt~=1.15.0
Cython~=3.0.0
setuptools~=65.5.0
patsy~=0.5.3
regex~=2022.8.17
mpmath~=1.2.1
sympy~=1.11.1
decorator~=5.1.1
astunparse~=1.6.3
tensorboard~=2.10.1
transformers~=4.21.1
urllib3~=1.26.12
Pygments~=2.16.1
cachetools~=5.3.1
certifi~=2022.6.15
MarkupSafe~=2.1.2
joblib~=1.3.1
python-dateutil~=2.8.2
packaging~=21.3
backcall~=0.2.0
pickleshare~=0.7.5
exceptiongroup~=1.1.3
wcwidth~=0.2.6
threadpoolctl~=3.2.0
Markdown~=3.4.4
networkx~=3.0
oauthlib~=3.2.2
Pyrogram~=2.0.41
idna~=3.3
asttokens~=2.4.0
contourpy~=1.1.0
executing~=1.2.0
fonttools~=4.42.0
pyparsing~=3.0.9
termcolor~=2.3.0
cycler~=0.11.0
kiwisolver~=1.4.4
flatbuffers~=23.5.26
tokenizers~=0.12.1
MetaTrader5~=5.0.45
retrying~=1.3.4
flask~=3.0.0
numexpr~=2.8.7
ta~=0.11.0